cleaned up FASTQC and other minor edits
Simplified FASTQC so that we can run it twice, once for raw fastq and
once for trimmed data.
Adjusted pipeline for FASTQC adjustments.
Minor documentation fixes and cleaning up of code.

+++ ngs_BLASTDB.sh
@@ -117,12 +117,10 @@ ngsCmd_BLASTDB() {
 	fi
 	
 	prnCmd "makeblastdb -in raw.fa -dbtype nucl -out $SAMPLE -title \"$SAMPLE\""
-	if ! $DEBUG; then 
-		makeblastdb -in raw.fa -dbtype nucl -out $SAMPLE -title "$SAMPLE"
-	fi
+	if ! $DEBUG; then makeblastdb -in raw.fa -dbtype nucl -out $SAMPLE -title "$SAMPLE"; fi
 	
 	prnCmd "rm raw.fa"
-	rm raw.fa
+	if ! $DEBUG; then rm raw.fa; fi
 	
 	# return to proper directory and restore $JOURNAL
 	prnCmd "cd $CUR_DIR"
